requirements.txt
astropy
matplotlib
pwntools
numpy
